Feature/update event endpoint

+++ server/src/v1/services/event.service.ts
@@ -2,6 +2,7 @@ import eventRepository from '../repositories/event.repository';
 import { AddEventInterface } from '../interface/event';
 import { Prisma } from '@prisma/client';
 import { NODE_ENV } from '../../constants/app.constants';
+import { NotFoundError, ForbiddenError } from '@/utils/customErrors';
 
 const addEvent = async (event_data: AddEventInterface) => {
   try {
@@ -27,13 +28,39 @@ const addEvent = async (event_data: AddEventInterface) => {
   }
 };
 
-const deleteEvent = async (eventId: string) => {
-  return await eventRepository.deleteEvent(eventId);
+const deleteEvent = async (
+  eventId: string,
+  created_by: string
+): Promise<boolean> => {
+  const event = await eventRepository.getEventDetails(eventId);
+
+  if (!event) {
+    throw new NotFoundError('Event not found');
+  }
+
+  if (event.created_by !== created_by) {
+    throw new ForbiddenError('You are not authorized to delete this event');
+  }
+
+  await eventRepository.deleteEvent(eventId);
+  return true;
+};
+
+const updateEvent = async (eventId: string, event_data: AddEventInterface) => {
+  const event = await eventRepository.getEventDetails(eventId);
+  if (!event) {
+    throw new NotFoundError('Event not found');
+  }
+  if (event.created_by !== event_data.created_by) {
+    throw new ForbiddenError('You are not authorized to update this event');
+  }
+  return await eventRepository.updateEvent(eventId, event_data);
 };
 
 const eventServices = {
   addEvent,
   deleteEvent,
+  updateEvent,
 };
 
 export default eventServices;
